% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_nbem.R
\name{create_test_data}
\alias{create_test_data}
\title{Create test data for BNBEM}
\usage{
create_test_data(
  n_labeled = 50,
  n_unlabeled = 100,
  n_words = 100,
  n_classes = 3,
  class_separation = 2,
  seed = 42
)
}
\arguments{
\item{n_labeled}{Number of labeled documents per class}

\item{n_unlabeled}{Number of unlabeled documents}

\item{n_words}{Number of words in vocabulary}

\item{n_classes}{Number of classes}

\item{class_separation}{Numeric value controlling class separation (higher = more separable)}

\item{seed}{Random seed (default: 42)}
}
\value{
List containing:
  - labeled_data: List with documents and labels
  - unlabeled_data: List of unlabeled documents
  - true_params: List of true parameters used to generate data
}
\description{
Generate artificial Bernoulli distribution vectors correlated with classes
}
