% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_nbem.R
\name{predict.BNBEM}
\alias{predict.BNBEM}
\title{Predict class for new documents}
\usage{
\method{predict}{BNBEM}(object, new_data, return_probs = FALSE)
}
\arguments{
\item{object}{BNBEM model}

\item{new_data}{List of new document word stems}

\item{return_probs}{Logical indicating if class probabilities should be returned}
}
\value{
List containing:
  - predictions: Character vector of predicted classes
  - probabilities: Matrix of class probabilities (if return_probs = TRUE)
}
\description{
Predict class for new documents
}
