% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_nbem.R
\name{print.BNBEM}
\alias{print.BNBEM}
\title{Print method for BNBEM objects}
\usage{
\method{print}{BNBEM}(x, ...)
}
\arguments{
\item{x}{BNBEM object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for BNBEM objects
}
