% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_nbem.R
\name{summary.BNBEM}
\alias{summary.BNBEM}
\title{Summary method for BNBEM objects}
\usage{
\method{summary}{BNBEM}(object, ...)
}
\arguments{
\item{object}{BNBEM object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for BNBEM objects
}
